# documentation/static/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 'FS PBX API'
  description: ''
  version: v1
servers:
  -
    url: 'https://{host}'
    description: 'FS PBX API server'
    variables:
      host:
        default: YOUR-FSPBX-HOST
        description: 'Hostname where FS PBX is running (example: fspbx.your-domain.net)'
paths:
  /api/v1/domains:
    get:
      summary: 'List domains'
      operationId: listDomains
      description: "Returns only the domains the caller is allowed to access.\n\nAccess rules:\n- If the user has the `domain_all` permission, all domains are returned.\n- If the user has any assigned domain groups or individual domains, only those domains are returned\n  (even if the user's own domain is not included).\n- If nothing is assigned, only the user's own domain is returned.\n\nPagination (cursor-based):\n- Both `limit` and `starting_after` are optional.\n- If `limit` is not provided, it defaults to 25.\n- If `starting_after` is not provided, results start from the beginning.\n- If `has_more` is true, request the next page by passing `starting_after`\n  equal to the last item's `domain_uuid` from the previous response.\n\nExamples:\n- First page: `GET /api/v1/domains`\n- Next page:  `GET /api/v1/domains?starting_after={last_domain_uuid}`\n- Custom size: `GET /api/v1/domains?limit=50`"
      parameters:
        -
          in: query
          name: limit
          description: 'Optional. Number of results to return (min 1, max 100). Defaults to 25.'
          example: 25
          required: false
          schema:
            type: integer
            description: 'Optional. Number of results to return (min 1, max 100). Defaults to 25.'
            example: 25
        -
          in: query
          name: starting_after
          description: 'Optional. Return results after this domain UUID (cursor).'
          example: 7d58342b-2b29-4dcf-92d6-e9a9e002a4e5
          required: false
          schema:
            type: string
            description: 'Optional. Return results after this domain UUID (cursor).'
            example: 7d58342b-2b29-4dcf-92d6-e9a9e002a4e5
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                example:
                  object: list
                  url: /api/v1/domains
                  has_more: true
                  data:
                    -
                      domain_uuid: 4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b
                      object: domain
                      domain_name: 10001.fspbx.com
                      domain_enabled: true
                      domain_description: 'BluePeak Solutions'
                properties:
                  object:
                    type: string
                    example: list
                  url:
                    type: string
                    example: /api/v1/domains
                  has_more:
                    type: boolean
                    example: true
                  data:
                    type: array
                    example:
                      -
                        domain_uuid: 4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b
                        object: domain
                        domain_name: 10001.fspbx.com
                        domain_enabled: true
                        domain_description: 'BluePeak Solutions'
                    items:
                      type: object
                      properties:
                        domain_uuid:
                          type: string
                          example: 4018f7a3-8e0a-47bb-9f4f-04b1313e0e1b
                        object:
                          type: string
                          example: domain
                        domain_name:
                          type: string
                          example: 10001.fspbx.com
                        domain_enabled:
                          type: boolean
                          example: true
                        domain_description:
                          type: string
                          example: 'BluePeak Solutions'
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  error:
                    type: authentication_error
                    message: Unauthenticated.
                    code: unauthenticated
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                        example: authentication_error
                      message:
                        type: string
                        example: Unauthenticated.
                      code:
                        type: string
                        example: unauthenticated
      tags:
        - Domains
    post:
      summary: 'Create a domain'
      operationId: createADomain
      description: "Creates a new domain.\n\nNotes:\n- `domain_enabled` is optional and defaults to `true`.\n- `domain_name` is normalized to lowercase and trimmed."
      parameters: []
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                type: object
                example:
                  domain_uuid: 9b6a4aa2-2b4f-4c5a-b4bb-1f6b2a9b9b01
                  object: domain
                  domain_name: 10005.fspbx.com
                  domain_enabled: true
                  domain_description: 'BluePeak Solutions'
                properties:
                  domain_uuid:
                    type: string
                    example: 9b6a4aa2-2b4f-4c5a-b4bb-1f6b2a9b9b01
                  object:
                    type: string
                    example: domain
                  domain_name:
                    type: string
                    example: 10005.fspbx.com
                  domain_enabled:
                    type: boolean
                    example: true
                  domain_description:
                    type: string
                    example: 'BluePeak Solutions'
        400:
          description: 'Domain name already exists'
          content:
            application/json:
              schema:
                type: object
                example:
                  error:
                    type: invalid_request_error
                    message: 'Domain name already exists.'
                    code: domain_name_exists
                    param: domain_name
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                        example: invalid_request_error
                      message:
                        type: string
                        example: 'Domain name already exists.'
                      code:
                        type: string
                        example: domain_name_exists
                      param:
                        type: string
                        example: domain_name
        401:
          description: Unauthenticated
          content:
            application/json:
              schema:
                type: object
                example:
                  error:
                    type: authentication_error
                    message: Unauthenticated.
                    code: unauthenticated
                properties:
                  error:
                    type: object
                    properties:
                      type:
                        type: string
                        example: authentication_error
                      message:
                        type: string
                        example: Unauthenticated.
                      code:
                        type: string
                        example: unauthenticated
      tags:
        - Domains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                domain_description:
                  type: string
                  description: 'A human-friendly label for the domain. Must not be greater than 255 characters.'
                  example: 'BluePeak Solutions'
                domain_name:
                  type: string
                  description: 'The domain name (lowercase). Must not be greater than 255 characters.'
                  example: 10001.fspbx.com
                domain_enabled:
                  type: boolean
                  description: 'Whether the domain is enabled.'
                  example: false
              required:
                - domain_description
                - domain_name
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []